Website makeover with hand-made CSS (#42)
Redo the entire design of the website using only vanilla CSS. Removed
all javascript and Bootstrap. Responsive columns are achieved through
flexbox and setting a minwidth on one of the columns. The CSS is much
simpler and easier to maintain. Updated some parts of the lab manual.
Got rid of some pages to make the website a bit more compact. Took a lot
of inspiration from these writings by Andy Bell:

* https://andy-bell.co.uk/cube-css/
* https://andy-bell.co.uk/a-modern-css-reset/
* https://set.studio/some-simple-ways-to-make-content-look-good/